+++ app/src/androidTest/java/com/pyildizt/sony/englishword/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.sony.englishword;
+package com.pyildizt.sony.englishword;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
