+++ app/src/main/java/com/pyildizt/sony/englishword/Splash.java
@@ -1,4 +1,4 @@
-package com.example.sony.englishword;
+package com.pyildizt.sony.englishword;
 
 import android.app.Activity;
 import android.content.Intent;
