+++ app/src/main/java/com/pyildizt/sony/englishword/firstLevel.java
@@ -1,10 +1,9 @@
-package com.example.sony.englishword;
+package com.pyildizt.sony.englishword;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.Button;
 
 public class firstLevel extends AppCompatActivity {
