+++ app/src/main/java/com/pyildizt/sony/englishword/meanFirst.java
@@ -1,4 +1,4 @@
-package com.example.sony.englishword;
+package com.pyildizt.sony.englishword;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
