+++ app/src/main/java/com/pyildizt/sony/englishword/secondLevel.java
@@ -1,4 +1,4 @@
-package com.example.sony.englishword;
+package com.pyildizt.sony.englishword;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
