+++ app/src/main/java/com/pyildizt/sony/englishword/testFirst.java
@@ -1,4 +1,4 @@
-package com.example.sony.englishword;
+package com.pyildizt.sony.englishword;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
@@ -16,23 +16,22 @@ import com.google.firebase.database.ValueEventListener;
 
 import java.util.Random;
 
-public class testSecond extends AppCompatActivity {
-
-   Boolean dogru=true;
-    int sayac=0,dongu,sayac1=31,rasgele=0,deneme=0,sinir=0,Bitis=0;
+public class testFirst extends AppCompatActivity {
+    Boolean dogru=true;
+    int sayac=0,dongu,sayac1=1,rasgele=0,deneme=0,sinir=2,Bitis=0;
     DatabaseReference oku;
     String kelime,kelimeK,kelimeK1,kelimeK2,kelimeK3;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_test_second);
+        setContentView(R.layout.activity_test_first);
         final Button tst1=(Button)findViewById(R.id.test1);
         final Button tst2=(Button)findViewById(R.id.test2);
         final Button tst3=(Button)findViewById(R.id.test3);
         final Button tst4=(Button)findViewById(R.id.test4);
         final TextView txt= (TextView)findViewById(R.id.testview);
 
-        for (dongu=31;dongu<35;dongu++) {
+        for (dongu=1;dongu<5;dongu++) {
             oku = FirebaseDatabase.getInstance().getReference().child("users").child(String.valueOf(dongu));
             ValueEventListener dinle = new ValueEventListener() {
                 @Override
@@ -73,7 +72,7 @@ public class testSecond extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 if (Bitis > 10){
-                    Intent ıntent = new Intent(testSecond.this, MainActivity.class);
+                    Intent ıntent = new Intent(testFirst.this, MainActivity.class);
                     startActivity(ıntent);
                 }
                 if(kelimeK.equals(txt.getText().toString()))
@@ -217,7 +216,7 @@ public class testSecond extends AppCompatActivity {
             public void onClick(View v) {
                 if(Bitis>10)
                 {
-                    Intent ıntent=new Intent(testSecond.this,MainActivity.class);
+                    Intent ıntent=new Intent(testFirst.this,MainActivity.class);
                     startActivity(ıntent);
                 }
                 if(kelimeK1.equals(txt.getText().toString()))
@@ -364,7 +363,7 @@ public class testSecond extends AppCompatActivity {
             public void onClick(View v) {
                 if(Bitis>10)
                 {
-                    Intent ıntent=new Intent(testSecond.this,MainActivity.class);
+                    Intent ıntent=new Intent(testFirst.this,MainActivity.class);
                     startActivity(ıntent);
                 }
                 if(kelimeK2.equals(txt.getText().toString()))
@@ -509,7 +508,7 @@ public class testSecond extends AppCompatActivity {
             public void onClick(View v) {
                 if(Bitis>10)
                 {
-                    Intent ıntent=new Intent(testSecond.this,MainActivity.class);
+                    Intent ıntent=new Intent(testFirst.this,MainActivity.class);
                     startActivity(ıntent);
                 }
                 if(kelimeK3.equals(txt.getText().toString()))
@@ -652,4 +651,3 @@ public class testSecond extends AppCompatActivity {
 
     }
 }
-
