+++ app/src/main/java/com/pyildizt/sony/englishword/yazFirst.java
@@ -1,4 +1,4 @@
-package com.example.sony.englishword;
+package com.pyildizt.sony.englishword;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -15,21 +15,20 @@ import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
 
-public class yazSecond extends AppCompatActivity {
+public class yazFirst extends AppCompatActivity {
 
     DatabaseReference oku,okuu;
     TextView t1;
     EditText e1;
     Button b1;
     String Englih="";
-    int sayac=30,id=0;
+    int sayac=0,id=0;
     boolean dogru=true;
-
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_yaz_second);
+        setContentView(R.layout.activity_yaz_first);
+
 
         t1=(TextView)findViewById(R.id.t1);
         e1=(EditText)findViewById(R.id.e1);
@@ -40,14 +39,15 @@ public class yazSecond extends AppCompatActivity {
         b1.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v){
-                if(sayac>45)
+                if(sayac>15)
                 {
-                    Intent intent=new Intent(yazSecond.this,MainActivity.class);
+                    Intent intent=new Intent(yazFirst.this,MainActivity.class);
                     startActivity(intent);
                 }
 
                 b1.setText("Sonraki");
                 e1.setVisibility(View.VISIBLE);
+
 
                 if(Englih.equals(e1.getText().toString())){
                     dogru=true;
