+++ app/src/main/res/layout/activity_mean_first2.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/backgroundnew"
-    tools:context="com.example.sony.englishword.meanFirst"
+    tools:context="com.pyildizt.sony.englishword.meanFirst"
     tools:layout_editor_absoluteY="81dp">
 
     <TextView
