+++ app/src/main/res/layout/activity_mean_second.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/backgroundnew"
-    tools:context="com.example.sony.englishword.meanSecond"
+    tools:context="com.pyildizt.sony.englishword.meanSecond"
     tools:layout_editor_absoluteY="81dp">
 
     <TextView
