+++ app/src/main/res/layout/activity_mean_third.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/backgroundnew"
-    tools:context="com.example.sony.englishword.meanThird">
+    tools:context="com.pyildizt.sony.englishword.meanThird">
 
     <Button
         android:id="@+id/ileri3"
