+++ app/src/main/res/layout/activity_test_first.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/backgroundnew"
-    tools:context="com.example.sony.englishword.testFirst"
+    tools:context="com.pyildizt.sony.englishword.testFirst"
     tools:layout_editor_absoluteY="81dp">
 
     <Button
