+++ app/src/main/res/layout/activity_test_second.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/backgroundnew"
-    tools:context="com.example.sony.englishword.testSecond">
+    tools:context="com.pyildizt.sony.englishword.testSecond">
 
     <Button
         android:id="@+id/test1"
