+++ app/src/main/res/layout/activity_third_level.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/backgroundnew"
-    tools:context="com.example.sony.englishword.firstLevel">
+    tools:context="com.pyildizt.sony.englishword.firstLevel">
 
     <TextView
         android:id="@+id/textView"
