+++ app/src/main/res/layout/activity_veri_ekle.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.sony.englishword.VeriEkle">
+    tools:context="com.pyildizt.sony.englishword.VeriEkle">
 
     <EditText
         android:id="@+id/tr"
