+++ app/src/main/res/layout/activity_yaz_first.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/backgroundnew"
-    tools:context="com.example.sony.englishword.yazFirst"
+    tools:context="com.pyildizt.sony.englishword.yazFirst"
     tools:layout_editor_absoluteY="81dp">
 
     <TextView
