+++ app/src/main/res/layout/activity_yaz_third.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/backgroundnew"
-    tools:context="com.example.sony.englishword.yazThird">
+    tools:context="com.pyildizt.sony.englishword.yazThird">
     <TextView
         android:id="@+id/t1"
         android:layout_width="132dp"
